WindBorne_Finance_Assignment/requirements.txt
requests
pandas
psycopg2-binary
sqlalchemy
streamlit
python-dotenv
plotly
